Rename SignOutButtonComponent to RightHeaderButtons
