+++ app/components/RightHeaderButtons.js
@@ -2,6 +2,7 @@
 import React, { Component } from 'react';
 import { Actions } from 'react-native-router-flux';
 import { connect } from 'react-redux';
+import { View } from 'react-native';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { authLogout } from '../../src/actions/index';
@@ -9,7 +10,7 @@ import ViewRow from '../base_components/ViewRow';
 import RippleIcon from '../base_components/RippleIcon';
 import PrimaryText from '../base_components/PrimaryText';
 
-class SignOutButton extends Component {
+class RightHeaderButtons extends Component {
   componentWillReceiveProps(nexProps, nextContext) {
     const { loginMessage } = nexProps;
     if (!loginMessage || !loginMessage.token) {
@@ -48,26 +49,38 @@ class SignOutButton extends Component {
           size={20}
           onPress={() => Actions.cartScreen()}
         >
-          <PrimaryText style={{
-            position: 'absolute',
-            top: 0,
-            right: 25,
-          }}
+          <View
+            style={{
+              position: 'absolute',
+              right: 25,
+              top: 0,
+              width: 25,
+              height: 25,
+              justifyContent: 'center',
+              borderRadius: 100,
+              backgroundColor: '#888',
+            }}
           >
-            {this.props.cartData.length}
-          </PrimaryText>
+            <PrimaryText style={{
+              color: '#fff',
+              fontSize: 14,
+            }}
+            >
+              {this.props.cartData.length}
+            </PrimaryText>
+          </View>
         </RippleIcon>
       </ViewRow>
     );
   }
 }
 
-SignOutButton.defaultProps = {
+RightHeaderButtons.defaultProps = {
   loginMessage: {},
   cartData: [],
 };
 
-SignOutButton.propTypes = {
+RightHeaderButtons.propTypes = {
   loginMessage: PropTypes.object,
   cartData: PropTypes.array,
   authLogout: PropTypes.func.isRequired,
@@ -88,4 +101,4 @@ function initMapDispatchToProps(dispatch) {
 }
 
 
-export default connect(initMapStateToProps, initMapDispatchToProps)(SignOutButton);
+export default connect(initMapStateToProps, initMapDispatchToProps)(RightHeaderButtons);
